# Introduce eager rent collection