--- a/runtime/src/accounts_index.rs
+++ b/runtime/src/accounts_index.rs
@@ -22,8 +22,7 @@ pub struct AccountsIndex<T> {
 }
 
 impl<'a, T: 'a + Clone> AccountsIndex<T> {
-    /// call func with every pubkey and index visible from a given set of ancestors
-    pub fn do_scan_accounts<F, I>(&self, ancestors: &Ancestors, mut func: F, iter: I)
+    fn do_scan_accounts<F, I>(&self, ancestors: &Ancestors, mut func: F, iter: I)
     where
         F: FnMut(&Pubkey, (&T, Slot)) -> (),
         I: Iterator<Item = (&'a Pubkey, &'a AccountMapEntry<T>)>,
@@ -36,6 +35,7 @@ impl<'a, T: 'a + Clone> AccountsIndex<T> {
         }
     }
 
+    /// call func with every pubkey and index visible from a given set of ancestors
     pub fn scan_accounts<F>(&self, ancestors: &Ancestors, func: F)
     where
         F: FnMut(&Pubkey, (&T, Slot)) -> (),
@@ -43,6 +43,7 @@ impl<'a, T: 'a + Clone> AccountsIndex<T> {
         self.do_scan_accounts(ancestors, func, self.account_maps.iter());
     }
 
+    /// call func with every pubkey and index visible from a given set of ancestors with range
     pub fn range_scan_accounts<F, R>(&self, ancestors: &Ancestors, range: R, func: F)
     where
         F: FnMut(&Pubkey, (&T, Slot)) -> (),