#### Problem

rent-due (!= rent-exempt) accounts are never garbage-collected unless updated (i.e. lazy collection), which accumulates over time and can be a DoS attack vector.

Also, the current rent collection mechanism has some gaming factor due to being lazy. (noted below)

Even more, there is [some UX issue for the lazy rent collection](https://github.com/solana-labs/solana/issues/7413#issuecomment-615136356).

#### Summary of Changes

- scan the whole pubkey domain space over an epoch (or 2 day (if epoch is less than 2 day and not warmed-up); ~not implemented yet~ _implemented_) for each slot progress, while updating all accounts.
  - (cons) **`AccountIndex` was switched from `HashMap` to `BTreeMap` for range queries to realize that**
- (pro) thus, maximum number of `AppendVec` is now bound to the number of slots of given epoch, combined with the recently-introduced per-slot shrinking #9219.
  - Also, this fixes unbound increase of `AccountsIndex.roots`.
- as a side effect, the rent collection schedule is now universally consistent: all existing accounts and newly-created accounts both alike are collected for the next epoch for each epoch exactly once.
  - (pro) technically, we can remove `account.rent_epoch` as a bonus. (`ACCOUNT_STORAGE_OVERHEAD`: 128 -> 120, -6.25% for IO bandwidth)
- the iteration algorithm intentionally designed to be deterministic as a requirement because rent collection generally affects account delta bank hash, distribution, and capitalization. Also it's stete-less as desired to avoid ABI breakage.
- (pro) the system load is well spread across an epoch because we can assume vast of pubkey space is uniformly distributed.
- (pro) this doesn't introduce ABI break; only an epoch-triggered enactment `if` guard is needed.

##### minor problem of current rent colletion

_Also I found that the lazy rent collection is problematic because delayed rent collection doesn't account for past rent fees were different or not._
That means the current calculation is based on the latest rent fee multiple of epochs. it's not the integral of past actual (dynamic) rent fees over epochs.
This means it introduces some gaming much like recently removed `redeem-vote-credits` (#7916):
A validator or an account owner might send 0 lamports to rent-due accounts to collect/save rent when the at-the-time dynamic rent is more profitable.
Although, the dynamic rent isn't still implemented...

##### (cons) new dos vector

An attacker can send lamports to arbitrary address if prepared to burn them so we're vulnerable to a new attack vector of targeted DoS (send so many pubkeys under some subrange corresponding to victim leader's slot), given publicly available deterministic next epoch's leader schedule.
But, I think that's tolerable given that there is already easier attack with similar threat profile: just flood the victim leader by burst TXes for some specific slot.
Still, If forced to mitigate, I think cloning the whole `Pubkey` set is required and then `.iter().chunk(.len() / SLOT_COUNT)`-ing on it.

#### Alternatives:

- scan whole `account_index` and collect rent when new epoch begins
  - (pro) most straightforward to reason and to implement.
  - (cons) too much sudden system load at the start of new epoch
- upfront `clone()` of `account_index.keys()` at each epoch boundary and process `.chunk()` of it for each slot like this PR.
  - (pro) no need to resort to `BTreeMap`.
  - (cons) moderate cluster hiccup and ABI break on Bank/Snapshot; also must include that into trust chain in some way.
- lookup by slot/slot range:
  - (cons) account updates aren't uniformly distributed across the slots in general
- slot coalescence (WIP: #9319):
  -  (cons) as pointed, single `AppendVec` for several slots showed performance degradation in the past. Also, this solution still doesn't solve stale rent-due accounts and gaming problem.
- stale slot/AppendVec eviction/unmap by LRU and on-demand mmap after eviction:
  - (cons) still too many unbound-number-of small AppendVec in the snapshot
- cold-store stale accounts:
  - (cons) too much effort to implement; could implemented later if there is so much account storage need. hampers the accounts index under huge number of active pubkeys.
  - (cons) also, we need to revise economic design/fee for executing TXes on accounts in the cold-storage; otherwise, economically-efficient DoS attack is possible
- maintain a `HashSet<Pubkey>` for rent-due accounts, separately:
   - (cons) can be another DoS vector when those account set gets large and we need a deterministic iteration algorithm.

(Also, I've lightly researched the rent mechanism of other blockchain projects and found there is no better solution. ;)

#### todo (once the direction is confirmed)

- ~[ ] update docs for rents~ (let's do with another future pr, this pr is already big).
- [x] hard fork logic
- [x] write/fix tests
- [x] refactor the copy and paste
- [x] check perf. degradation of using `BTreeMap` over `HashMap` (doing right now)
- [x] warmup epochs & rent? what to do? (DONE: introduced constant cycle collection).

Fixes #9383