@sakridge 2020-04-16
Cool. I didn't review the code in-depth yet, but from the description and scanning it, it seems like a really good direction.