@ryoqun 2020-04-23
metrics is now working:

![image](https://user-images.githubusercontent.com/117807/80058767-2c523e80-8565-11ea-8719-a449accf5862.png)