@ryoqun 2020-04-27
At least, the performance degradation should be minimal for eager rent collection itself. we still need to measure the effect of switching from `HashMap` to `BTreeMap`, though.

![image](https://user-images.githubusercontent.com/117807/80383005-fbb03300-88dd-11ea-9fba-313b8db21c0f.png)