@ryoqun 2020-04-28
> hmm. I couldn't conclude any performance observation this time. It seems that I need to tune some benchmark parameters to bet stable results...

may this situation is affected by #9739