@sakridge 2020-05-05
Seems pretty good to me. So it collects an 8-byte prefix of pubkeys each slot?