@ryoqun 2020-05-11
> Seems pretty good to me. So it collects an 8-byte prefix of pubkeys each slot?

@sakridge Yeah, that's correct. Strictly speaking, it collects a range of 8-byte prefix of pubkeys each slot, though.