@ryoqun 2020-05-11
> @ryoqun - are you just waiting for review on this PR or are you still working out some things first?

@mvines At the moment, there is little work left. It seems that [review is done](https://github.com/solana-labs/solana/pull/9527#issuecomment-626576089) (yay!) and some perf degradation check reporting is required. (ran bunch of perfs, which, it seems there are negligible impact at first glance).