@ryoqun 2020-05-11
I've prepared pdfs...

[base-without-btreemap-and-eager-rent-collection.pdf](https://github.com/solana-labs/solana/files/4608757/base-without-btreemap-and-eager-rent-collection.pdf)
[variant-1-with-btreemap-and-without-eager-rent-collection.pdf](https://github.com/solana-labs/solana/files/4608758/variant-1-with-btreemap-and-without-eager-rent-collection.pdf)
[variant-2-with-btreemap-and-eager-rent-collection.pdf](https://github.com/solana-labs/solana/files/4608759/variant-2-with-btreemap-and-eager-rent-collection.pdf)


I'll lightly comment on these with regard to my view of perf. impact.