@ryoqun 2020-05-12
### conclution

After careful review of perf. results, I don't see any integration-level perf. degradation.

### the methology.

I ran base bench runs twice before and after the whole benchmarking process.
In between in time, I ran the **variant1** where only `HashTree` => `BTreeMap` change is done (0b9482d69a3f5c30d8895dc7d4a1de4a3b59165f) and the **variant2** where the whole eager rent collection stuff is enabled (9f19166fccdaca428385dee9289b94ad9fb674a0). The base was at 754c65c066c12542fcd8580ed8118d82c18f9a48, which is the immediate parent commit of the variant1:

- [base-without-btreemap-and-eager-rent-collection.pdf](https://github.com/solana-labs/solana/files/4608757/base-without-btreemap-and-eager-rent-collection.pdf)
- [variant-1-with-btreemap-and-without-eager-rent-collection.pdf](https://github.com/solana-labs/solana/files/4608758/variant-1-with-btreemap-and-without-eager-rent-collection.pdf)
- [variant-2-with-btreemap-and-eager-rent-collection.pdf](https://github.com/solana-labs/solana/files/4608759/variant-2-with-btreemap-and-eager-rent-collection.pdf)



You can see that the eager rent collection is actually enabled in non-empty 'eager rent collection' panel in the grafana pdf of variant2. From that chart, you can see approximately 4 accounts are constantly rent-collected for each slot during the benchmark, resulting almost 0-ms collection time. (yeah, I should measure in us...)

(Also, ignore "Software Version", it says all the same `3ee702a92`, maybe HEAD of the master at the time)

### observation

I can't see any notable chart changes, in my eyes, these charts looks too similar to differentiate.
the only unrelated note is that there is no work of `bank-forks_set_root` before the `bench-tps` starts in the variant2's `time spent in squashing` panel for unknown reason. 

As for the numbers, these are so similar too:


||base (2-runs)|variant1|variant2|
|-|-|-|-|
|mean tps| 16712-16857 |16821|16918|
|mean conf. time| 1.042s-1.062s|1.051s|1.096s|
|median conf. time| 908ms-918ms|912ms|921ms|
|min conf. time| 169ms-173ms|171ms|173ms|
|max conf. time| 3.95s-3.97s|4.60s|4.33s|
|99 percentile conf. time| 2.481s-2.634s |2.683s|2.611s|