@ryoqun 2020-05-12
Also, I've run micro-benchmarks with release build on cpu-bound settings with fixed freq. as well:

As expected, BTreeMap is slower than HashMap around 15-30% (no `--clean`'s initlal account creation, 37.8s -> 53.7s; `--clean`'s clean 690ms -> 810ms, prepare: 6.1s -> 7.2s).
Unexpectedly and surprisingly, the sequencial read perf. is improved somewhat: 1.9s -> 1.7s 1.11x (no `--clean` mode's hashing time). This might offset some of the slowdown.

At this microscopic level, we can measure the difference, but I think these are small part in [the integration-level benchmark](https://github.com/solana-labs/solana/pull/9527#issuecomment-627210347). So, we couldn't measure the performance difference there.

Also, as for memory use, there was no notable difference:


### `solana-accounts-bench` (no `--clean`)


```
$ env time -v /tmp/solana-accounts-bench-with-hashmap --num_accounts 6000000 --num_slots 400
clean: false
Creating 6000000 accounts
created 6000000 accounts in 400 slots create accounts took 37.8s
hash: BEDjLurPVDuvoqzUAx9bFVtaCWRSGRZFLsgzuHspuW5t hash took 2.3s prepare took 26us
hash: 57Q4WU9Fa1n6p1fcmQRJfvEa3uh1oTEZReuakSweCbBk hash took 1.9s prepare took 21us
hash: 9HWsmhXZ3BkbqFgnWggLmxDapCp3xFfoy5ah3SCA9qyA hash took 1.9s prepare took 20us
hash: GoDemUt4JqhGfk91JqpUE3bsKWWyz6CPssn1wgSxC8Jj hash took 1.9s prepare took 21us
hash: nWXF97fCjcZWQXW26VMdhmCqdwtM1JxaDNWZta7jF25 hash took 2.0s prepare took 20us
hash: 6t8aRF5oA2dtL9tbsqEQzZJz7BwnSdBaWd6uRMKpkxj1 hash took 1.9s prepare took 22us
hash: 7zFRBvSMzcZrX1AhzgnaS71eEo5jmDXK7Xwfcf93pN5a hash took 1.9s prepare took 43us
hash: 3rq9swAqDD12xsJo8ZqsyGChMnHhffT7ic2JgjffPEp4 hash took 1.9s prepare took 21us
hash: Be4VHK4CZGRKufi4gUpD6ZBMfUq6i1axwYidyaYjzyQe hash took 2.0s prepare took 20us
hash: 6XdekPLrx7faFuHh98ifhbV8nNHytuKTuoMQUVkrDmh6 hash took 2.0s prepare took 22us
hash: F5gYJ2n6RdyHzhZcrv3rYo5mudBSFeGRPRi1YQCW6uxa hash took 2.0s prepare took 21us
hash: 8VGMumVk4KXv2e6VpEZDrFPMUgH1Sm1DW9NLG2oF22Mr hash took 1.9s prepare took 21us
hash: Epumgi8m9XbL6hVc4XhJ9u9XCMFrwbnnRpEJFXvUD2ve hash took 2.0s prepare took 22us
hash: AEkmVKVb93VjmYYxTQUmMTSXc5grnuYhBFzAc6XjaaD5 hash took 2.0s prepare took 22us
hash: BmywVfYjBqb62BtsqCYuxX1wpMPD8Gj32oN3MWcehHxd hash took 2.0s prepare took 22us
hash: 3efBfk1CiWpJt8jRAmYPpC4PJNgALVWA2eJdmpMfSnMb hash took 2.0s prepare took 20us
hash: 6TDv6NBNMF3Vyb1cV4KcrGGPSYWzCibpGuem9FsiDY2R hash took 2.0s prepare took 21us
hash: EZdXtq9eUPnhqzqD6o34A3pX6yTmNJseShiBMtRxfkoQ hash took 1.9s prepare took 21us
hash: 3cZAY1ztSttCLiPb5XPe7C1GbUbx24ZYBbBQwPPwpkUp hash took 2.0s prepare took 22us
hash: GQLdpi5fMeNRSWE5xRvN1cJ4m6Gb3nV2PfuoBBeE2TES hash took 2.0s prepare took 21us
        Command being timed: "/tmp/solana-accounts-bench-with-hashmap --num_accounts 6000000 --num_slots 400"
        User time (seconds): 318.41
        System time (seconds): 55.62
        Percent of CPU this job got: 454%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 1:22.29
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 10484456
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 4415784
        Voluntary context switches: 8346845
        Involuntary context switches: 6675
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

```
$ env time -v /tmp/solana-accounts-bench-with-btreemap --num_accounts 6000000 --num_slots 400
clean: false
Creating 6000000 accounts
created 6000000 accounts in 400 slots create accounts took 53.7s
hash: 5Rprx8mJZxo8atKt1Rd24y97BvGkEZWwXiSVevascBCL hash took 2.1s prepare took 30us
hash: ENDkKtUYLNEAmfnEMQA6TZ8pwB9eKYfVSvWtFs8Cj6xr hash took 1.9s prepare took 26us
hash: FqhjCNkxts1FQVgDdSEcVETaS3LT4nB8Zsr5aUzotF5e hash took 1.7s prepare took 28us
hash: CF87F3yL2jVURnZabN89fuf49jtRiT9XeJ9mTTDR17ni hash took 1.7s prepare took 43us
hash: 2suzVs6ctoHknmf34XpDfrAcXyodBcJArF84WUvEQKDV hash took 1.7s prepare took 24us
hash: 9FwiqUGYgfcof5JoUUvqGgTsQJA8CfE5deEFVAGE34K1 hash took 1.7s prepare took 26us
hash: FcXAyEKC61ttGW7goNDY1oFxRwid7m2pBgTxseRD9BBX hash took 1.7s prepare took 23us
hash: AKqVjiEpczXrnjTqrW3V8SfAtu5KrqxkdvvCiZLbQknA hash took 1.7s prepare took 24us
hash: 2cxENRuCjQfRX98raA8BXmSmcXrARMT43ADDAJayjhjo hash took 1.7s prepare took 24us
hash: AfZ1YG2Cb87yMhW2un6dS8VBZarpggYxGNuEBvxMyX9r hash took 1.7s prepare took 25us
hash: Dr45n5oZTjxtz2ivWtYxottP6fzWPF6qEFAvA1HFLfgz hash took 1.7s prepare took 26us
hash: GAumprF5Fsf6dE8DtqypbJEvAK6ScunCavSvD4ST76Hc hash took 1.7s prepare took 24us
hash: H6QVcAGvhg8PivkEXNib39kGHyXEz9JNcBa1w8WfvB4L hash took 1.7s prepare took 25us
hash: DyjKC1zmRn7yGU9NsCi79LjBk1UAhodykimV2pUzx46W hash took 1.7s prepare took 43us
hash: 3qnmYsUdEB6REd6ub2zce7TRJEnNXBsYDU4BtBcp6kdy hash took 1.7s prepare took 25us
hash: 8vE5Bus4JaBmy2a1F9vDyHe5AjJsNJKdw3brzBt9dyFg hash took 1.7s prepare took 46us
hash: 2NGPKfcVeYiSfJV4VLzyXyEee2yoFakS2tC78ULJFMNN hash took 1.7s prepare took 25us
hash: 9x84fiPtwHmhgricwW8t4GrQVNhjAdL2zYzGxVk8PvZg hash took 1.7s prepare took 24us
hash: EUNUGTUkKLCGLAjvLSgFeF4BYunxJryxiC1RJoqRRqyb hash took 1.7s prepare took 26us
hash: GS5ZjAbAYVQgdnbG1d7A5XfH6uKngwPyRq5QsigeZmny hash took 1.7s prepare took 25us
        Command being timed: "/tmp/solana-accounts-bench-with-btreemap --num_accounts 6000000 --num_slots 400"
        User time (seconds): 280.66
        System time (seconds): 61.85
        Percent of CPU this job got: 364%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 1:34.02
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 10258316
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 2899821
        Voluntary context switches: 10247861
        Involuntary context switches: 3930
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0

```

### `solana-accounts-bench --clean`

```
$ env time -v /tmp/solana-accounts-bench-with-hashmap --num_accounts 1000000 --num_slots 2 --clean
clean: true
Creating 1000000 accounts
created 1000000 accounts in 2 slots create accounts took 4.1s
clean took 155ms prepare took 5.8s
clean took 263ms prepare took 6.0s
clean took 693ms prepare took 6.0s
clean took 705ms prepare took 6.0s
clean took 693ms prepare took 6.0s
clean took 687ms prepare took 6.1s
clean took 695ms prepare took 6.1s
clean took 694ms prepare took 6.1s
clean took 673ms prepare took 6.0s
clean took 688ms prepare took 6.1s
clean took 667ms prepare took 6.1s
clean took 701ms prepare took 6.0s
clean took 683ms prepare took 6.0s
clean took 689ms prepare took 6.1s
clean took 671ms prepare took 6.1s
clean took 687ms prepare took 6.1s
clean took 688ms prepare took 6.1s
clean took 687ms prepare took 6.1s
clean took 686ms prepare took 6.1s
clean took 688ms prepare took 6.1s
        Command being timed: "/tmp/solana-accounts-bench-with-hashmap --num_accounts 1000000 --num_slots 2 --clean"
        User time (seconds): 165.87
        System time (seconds): 9.04
        Percent of CPU this job got: 125%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 2:18.97
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 7034028
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 2574319
        Voluntary context switches: 316042
        Involuntary context switches: 3151
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

```
$ env time -v /tmp/solana-accounts-bench-with-btreemap --num_accounts 1000000 --num_slots 2 --clean
clean: true
Creating 1000000 accounts
created 1000000 accounts in 2 slots create accounts took 7.6s
clean took 213ms prepare took 6.9s
clean took 336ms prepare took 7.1s
clean took 805ms prepare took 7.2s
clean took 805ms prepare took 7.2s
clean took 816ms prepare took 7.2s
clean took 808ms prepare took 7.2s
clean took 810ms prepare took 7.2s
clean took 829ms prepare took 7.2s
clean took 828ms prepare took 7.2s
clean took 807ms prepare took 7.2s
clean took 809ms prepare took 7.2s
clean took 817ms prepare took 7.2s
clean took 808ms prepare took 7.2s
clean took 804ms prepare took 7.2s
clean took 819ms prepare took 7.2s
clean took 810ms prepare took 7.2s
clean took 782ms prepare took 7.2s
clean took 816ms prepare took 7.2s
clean took 782ms prepare took 7.2s
clean took 823ms prepare took 7.2s
        Command being timed: "/tmp/solana-accounts-bench-with-btreemap --num_accounts 1000000 --num_slots 2 --clean"
        User time (seconds): 199.59
        System time (seconds): 12.31
        Percent of CPU this job got: 126%
        Elapsed (wall clock) time (h:mm:ss or m:ss): 2:47.06
        Average shared text size (kbytes): 0
        Average unshared data size (kbytes): 0
        Average stack size (kbytes): 0
        Average total size (kbytes): 0
        Maximum resident set size (kbytes): 7009912
        Average resident set size (kbytes): 0
        Major (requiring I/O) page faults: 0
        Minor (reclaiming a frame) page faults: 2626310
        Voluntary context switches: 1017412
        Involuntary context switches: 2798
        Swaps: 0
        File system inputs: 0
        File system outputs: 0
        Socket messages sent: 0
        Socket messages received: 0
        Signals delivered: 0
        Page size (bytes): 4096
        Exit status: 0
```

```
[5] pry(main)> 37.8/53.7
=> 0.7039106145251396
[7] pry(main)> 690.0/810
=> 0.8518518518518519
[8] pry(main)> 6.1/7.2
=> 0.8472222222222221
[9] pry(main)> 1.9/1.7
=> 1.1176470588235294
```